LayoutTests/media/media-source/media-source-real-seek-and-play.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-seek-and-play</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Tests sequential append of contiguous segments with real MP4 data.
    // Verifies buffered ranges merge correctly and endOfStream works.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        if (!MediaSource.isTypeSupported(MP4.VIDEO_TYPE)) {
            failTest(`${MP4.VIDEO_TYPE} is not supported`);
            return;
        }

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });

        // First segment: 4 seconds
        var media1 = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
            ]}]
        });

        // Second segment: 4 more seconds (contiguous)
        var media2 = MP4.mediaSegment({
            sequenceNumber: 2,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 4000, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
            ]}]
        });

        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        sourceBuffer.appendBuffer(media1);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.start(0)', 0, 0.01);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 4, 0.01);

        sourceBuffer.appendBuffer(media2);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.start(0)', 0, 0.01);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 8, 0.01);

        run('source.endOfStream()');
        testExpected('source.readyState', 'ended');
        testExpectedEqualWithTolerance('source.duration', 8, 0.01);

        endTest();
    }

    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
